move cache to api
